refactor(server): inline OAuth bearer resolver into require-auth middleware
resolveBearerUser was a 1-production-caller helper whose ResolverDeps shape
existed to let tests inject a substituted issuer and stub verifier. Two of
those tests covered paths that cannot fire in production: the issuer is
always derived from c.var.authBaseURL, so the wrong-issuer override branch
is unreachable; the "verifier was not called" assertion was an
implementation detail of a 4-line early-return.

Inlines the resolver into require-auth.ts as a private helper that reads
c.var.authBaseURL and c.var.db directly. Shrinks the auth/resource-boundary
file (101 lines) to auth/parse-bearer.ts (13 lines, just the header
utility). Migrates the three real-flow tests (valid token, wrong audience,
user removed mid-flight) to middleware integration tests that drive
requireBearerUser through a real Better Auth server.

ResolverDeps, VerifyOAuthAccessToken, and RequestOAuthEnv types all go
with the boundary.
