+++ packages/server/src/auth/parse-bearer.ts
@@ -1,0 +1,13 @@
+/**
+ * Extract the token from an HTTP `Authorization: Bearer <token>` header value.
+ * Case-insensitive on the scheme; trims surrounding whitespace; returns null
+ * for missing, empty, or non-bearer inputs.
+ *
+ * Shared between `single-credential` (well-formedness) and `require-auth`
+ * (authorization) so both layers agree on what counts as a bearer.
+ */
+export function parseBearer(value: string | null): string | null {
+	if (!value) return null;
+	const match = value.match(/^Bearer\s+(.+)$/i);
+	return match?.[1]?.trim() || null;
+}
